Refine section parsing into provision builder
