+++ tests/pdf_ingest/test_rule_extraction.py
@@ -21,15 +21,19 @@ def test_rule_extraction(monkeypatch, tmp_path):
 
     def fake_parse_sections(text):
         body = text.split("\n", 1)[1] if "\n" in text else text
-        node = types.SimpleNamespace(
-            text=body,
-            identifier="1",
-            heading="Heading",
-            node_type="section",
-            rule_tokens={"modality": "must", "conditions": [], "references": []},
-            children=[],
-        )
-        return [node]
+        return [
+            Provision(
+                text=body,
+                identifier="1",
+                heading="Heading",
+                node_type="section",
+                rule_tokens={
+                    "modality": "must",
+                    "conditions": [],
+                    "references": [],
+                },
+            )
+        ]
 
     monkeypatch.setattr(
         pdf_ingest,
